rosplan_starter_rcll: start planning when entering PRODUCTION phase
This little tool monitors the game state and triggers ROSPlan to start
planning when entering the production phase. A small configurable delay
allows to avoid race conditions, e.g., if orders to plan for have not
yet been received and updated by the planning system.

This is a tool to show the general idea, but it is obviously lacking in
that it plans exactly once and performs no monitoring whatsoever once
this has been done.